network_of_assistants/noa-moderator/pyproject.toml
[project]
name = "noa-moderator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "slim",
    "langchain>=0.3.23",
    "langchain-openai",
    "langchain-ollama>=0.3.1",
    "langchain-mistralai>=0.2.10",
    "pydantic-settings>=2.8.1",
    "ioa-observe-sdk>=1.0.12",
    "slim-bindings>=0.3.5",
]

[tool.uv.sources]
slim = { workspace = true }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
noa-moderator = "main:run"
